Added an empty settings screen and started work on a sleep System.
Added an empty settings screen that will be populated with features as they are required, eg. brightness. Started work on a sleep system, working on it I realized a callback system would make that and a lot of other systems much more efficient, that is what I will work on next time. Sleep system not working at the current moment.